RISC-V: Don't report an error until the link phase if suitable multilib isn't found.
When suitable multilib isn't found, an error is not reported until the
link period, which is inconsistent with the result of compiling option
`-print-multi-directory`. For example, when suitable multilib isn't
found, the return result of `-print-multi-directory` is the default
value '.', while the actual execution result is an error during the
link phase. This is not very reasonable.

I think the error should be reported in advance, so that the compilation
option `-print-multi-directory` will also report an error. The two
should be consistent, either reporting errors or using default values.

gcc/ChangeLog:

	* common/config/riscv/riscv-common.cc
	(riscv_compute_multilib): report an error if suitable multilib isn't found.
	(struct riscv_multi_lib_info_t): Delet.
	(riscv_multi_lib_check): Likewise.
	* config/riscv/elf.h (LIB_SPEC): Likewise.
	* config/riscv/riscv.h (riscv_multi_lib_check): Likewise.
	(EXTRA_SPEC_FUNCTIONS): Likewise.