parloops: Copy target and optimizations when creating a function clone
Hi,

This patch makes sure we copy over
DECL_FUNCTION_SPECIFIC_{TARGET,OPTIMIZATION} in parloops when creating
function clones.  This is required for SVE clones as we will need to
enable +sve for them, regardless of the current target options.
I don't actually need the 'OPTIMIZATION' for this patch, but it sounds
like a nice feature to have, so you can use pragmas to better control
options used in simd_clone generation.

gcc/ChangeLog:

         * tree-parloops.cc (create_loop_fn): Copy specific target and
optimization options
         when creating a function clone.

Is this OK for stage 1?