Speedup PTA solving for call constraint sets
With calls we now often get contraints like

  callarg = *callarg + UNKNOWN

or similar cases.  The important thing to note is that this
complex constraint changes the node solution itself, so when
solving the node is marked as changed immediately again.  When
that happens it's profitable to iterate that self-cycle immediately
so we maximize cache reuse and build up the successor graph quickly
to get better topological ordering and reduce the number of
iterations of the solving.

For a testcase derived from ceph this reduces the time spent in
PTA solving from 453s to 92s which is quite significant.

Bootstrap and regtest running on x86_64-unknown-linux-gnu.  For
the testcase I verified we create identical points-to solutions
before and after the change.  There are regression bugs complaining
about high PTA time (often only as part of overall slow compile),
I did not verify if this improves any of those but consider the
change a regression fix.

	* tree-ssa-structalias.cc (solve_graph): Immediately
	iterate self-cycles.