ifcvt: Lower bitfields only if suitable for scalar register [PR tree/109005]
This patch fixes the condition check for eligilibity of lowering bitfields,
where before we would check for non-BLKmode types, in the hope of excluding
unsuitable aggregate types, we now check directly the representative is
not an
aggregate type, i.e. suitable for a scalar register.

I tried adding the reduced testcase mentioned in the PR, but I couldn't
get the Ada testsuite to run, so could an Ada maintainer add the test
after verifying it runs properly?

OK for trunk?

gcc/ChangeLog:

	PR tree/109005
	* tree-if-conv.cc (get_bitfield_rep): Replace BLKmode check with
	aggregate type check.