tree: define tree_code_type in C++11/14 [PR108634]
Tested x86_64-pc-linux-gnu, OK for trunk?

-- 8< --

The r13-6577 change to use tree_code_type_tmpl in earlier C++ dialects broke
gdbhooks, which expects tree_code_type to always be available.  I considered
trying to make gdbhooks more robust, but it seemed simpler to define
tree_code_type as a reference to the template.  This still ends up with a
definition of the reference in each translation unit, but that's allowed by the
ODR because it always refers to the same entity, and is much smaller than
having the whole table in each TU.

	PR plugins/108634

gcc/ChangeLog:

	* tree-core.h (tree_code_type, tree_code_length):
	Define even without inline variable support.
	* tree.h (TREE_CODE_CLASS, TREE_CODE_LENGTH):
	Only one definition.